+++ frontend/package.json
@@ -62,9 +62,6 @@
     "@emotion/styled": "^11.8.1",
     "@headlessui/react": "^1.4.2",
     "@hookform/resolvers": "^2.8.8",
-    "@mui/icons-material": "^5.8.4",
-    "@mui/material": "^5.10.0",
-    "@mui/styles": "^5.4.4",
     "@next/bundle-analyzer": "14.1.4",
     "@sk-web-gui/core": "3.11.1",
     "@sk-web-gui/react": "1.4.0",
