+++ frontend/src/supportmanagement/components/support-errand/sidebar/sidebar-generic-notes.component.tsx
@@ -18,7 +18,7 @@ import {
   useConfirm,
   useSnackbar,
 } from '@sk-web-gui/react';
-import { GenericNote } from '@supportmanagement/components/notes-list/notes-list.component';
+import { ErrandNotesTabFormModel, GenericNote } from '@supportmanagement/interfaces/genericNote';
 import { getSupportErrandById } from '@supportmanagement/services/support-errand-service';
 import {
   SupportNote,
@@ -31,12 +31,6 @@ import dayjs from 'dayjs';
 import { Fragment, useEffect, useState } from 'react';
 import { UseFormReturn, useForm } from 'react-hook-form';
 import * as yup from 'yup';
-
-interface ErrandNotesTabFormModel {
-  id?: string;
-  partyId?: string;
-  text: string;
-}
 
 let formSchema = yup
   .object({
@@ -51,7 +45,7 @@ export const SidebarGenericNotes: React.FC<{
   label_singular: 'Kommentar' | 'Tjänsteanteckning';
   noteType: NoteType;
 }> = ({ label_plural, label_singular, noteType }) => {
-  const { user, supportErrand, setSupportErrand, administrators, municipalityId } = useAppContext();
+  const { supportErrand, setSupportErrand, administrators, municipalityId } = useAppContext();
   const [selectedNote, setSelectedNote] = useState<GenericNote>();
   const [notes, setNotes] = useState<SupportNote[]>([]);
   const [isLoading, setIsLoading] = useState(false);
