+++ frontend/src/supportmanagement/components/support-errand/tabs/messages/support-messages-tree.component.tsx
@@ -1,7 +1,6 @@
 import React, { Fragment, useState } from 'react';
 import { RenderedSupportMessage } from './rendered-support-message.component';
-import { Divider } from '@mui/material';
-import { Button, cx } from '@sk-web-gui/react';
+import { Button, cx, Divider } from '@sk-web-gui/react';
 import {
   MessageNode,
   countAllMessages,
@@ -104,7 +103,7 @@ const MessageTreeComponent: React.FC<MessageTreeProps> = ({
     <div className="my-lg">
       {nodes.map((node, idx) => (
         <Fragment key={`${idx}-${getId(node)}`}>
-          <Divider />
+          <Divider orientation="vertical" strong={true} />
           <MessageNodeComponent
             update={update}
             setRichText={setRichText}
